Move declared interfaces from ElementRef to Interfaces file
